16k: disable edgetpu SSMT and aux domains
This is needed to get the edgetpu kernel driver to handle the mailbox
allocations in order to have the kernel force the usage of only the even
mailbox IDs. We need to do this to support 16KB pages since only the
even client mailboxes fall on a 16KB page boundary. The alternative is
to add a new ioctl to allow userspace to query for the assigned mailbox
ID which would allow userspace to calculate the offset to it's mailbox
CSRs.

Bug: 270619308
Change-Id: I06c1bb6ca42f53c701530db8bb1e8cab9c7a7472
Signed-off-by: Will McVicker <willmcvicker@google.com>